scripts/requirements.txt
openai
turbopuffer
python-dotenv
